stable-diffusion-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
diffusers==0.25.0
transformers==4.36.0
torch>=2.0.0
torchvision>=0.15.0
accelerate==0.25.0
safetensors==0.4.1
Pillow>=10.0.0
